feat/3: chore - create container for login and register, move layout and navbar to common
