+++ apps/web/src/components/common/navigation-bar/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./navigation-bar";
