+++ apps/web/src/components/common/navigation-bar/navigation-bar.tsx
