+++ apps/web/src/components/login/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./login";
