--- apps/web/src/components/navigation-bar/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./navigation-bar.tsx";
